+++ CHANGELOG.md
@@ -64,9 +64,11 @@
 - Fixed a bug when use batched address. Since we were sorting the address during the forward pass, the output prediction 
   tags were not aligned with the supposed parsed address. We have removed the sorting, and now the results are more 
   aligned with our research.
+<<<<<<< HEAD
   
-## 0.4
+## dev
 
+- Added verbose flag to training and test base on the __init__ of address parser.
 - Added a feature to retrain our models with prediction tags dictionary different from the default one.
 - Added a feature to test our model with prediction tags dictionary different from the default one.
 - Added in-doc code examples.